type2문제 생성을 위한 dupDate테이블 추가 및 type2 문제 생성 로직 변경

+++ src/main/java/Project/OpenBook/Service/TopicService.java
@@ -1,23 +1,23 @@
 package Project.OpenBook.Service;
 
 import Project.OpenBook.CustomException;
-import Project.OpenBook.Domain.Category;
-import Project.OpenBook.Domain.Chapter;
-import Project.OpenBook.Domain.Topic;
-import Project.OpenBook.Dto.error.ErrorDto;
+import Project.OpenBook.Domain.*;
 import Project.OpenBook.Dto.topic.TopicDto;
 import Project.OpenBook.Repository.CategoryRepository;
 import Project.OpenBook.Repository.ChapterRepository;
+import Project.OpenBook.Repository.dupdate.DupDateRepository;
+import Project.OpenBook.Repository.choice.ChoiceRepository;
+import Project.OpenBook.Repository.description.DescriptionRepository;
 import Project.OpenBook.Repository.topic.TopicRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
 
+import java.util.ArrayList;
 import java.util.Arrays;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 import static Project.OpenBook.Constants.ErrorCode.*;
@@ -29,8 +29,13 @@ public class TopicService {
 
     private final TopicRepository topicRepository;
     private final CategoryRepository categoryRepository;
-
     private final ChapterRepository chapterRepository;
+
+    private final DupDateRepository dupDateRepository;
+
+    private final ChoiceRepository choiceRepository;
+
+    private final DescriptionRepository descriptionRepository;
 
     public TopicDto queryTopic(String topicTitle) {
         Topic topic = checkTopic(topicTitle);
@@ -43,11 +48,8 @@ public class TopicService {
     public Topic createTopic(TopicDto topicDto) {
 
         Category category = checkCategory(topicDto.getCategory());
-
         Chapter chapter = checkChapter(topicDto.getChapter());
-
         checkDupTopicTitle(topicDto.getTitle());
-
 
         Topic topic = Topic.builder()
                 .chapter(chapter)
@@ -59,8 +61,9 @@ public class TopicService {
                 .questionNum(0)
                 .choiceNum(0)
                 .build();
-
         topicRepository.save(topic);
+
+        addDupDates(topic);
         return topic;
     }
 
@@ -73,14 +76,36 @@ public class TopicService {
         int chapterNum = topicDto.getChapter();
         Chapter chapter = checkChapter(chapterNum);
 
+        boolean flag = false;
+
+        if(topic.getStartDate()!=topicDto.getStartDate() || topic.getEndDate() != topicDto.getEndDate()){
+            flag = true;
+        }
+
         topic.updateTopic(topicDto.getTitle(), topicDto.getStartDate(),topicDto.getEndDate(), topicDto.getDetail(),
                 chapter, category);
+
+        if (flag) {
+            List<DupDate> dupDateList = dupDateRepository.queryAllByTopic(topic.getTitle());
+            dupDateRepository.deleteAllInBatch(dupDateList);
+            addDupDates(topic);
+        }
 
         return topic;
     }
 
     public boolean deleteTopic(String topicTitle) {
         Topic topic = checkTopic(topicTitle);
+
+        List<Choice> choiceList = choiceRepository.queryChoiceByTopicTitle(topicTitle);
+        if (!choiceList.isEmpty()) {
+            throw new CustomException(TOPIC_HAS_CHOICE);
+        }
+
+        List<Description> descriptionList = descriptionRepository.findDescriptionsByTopic(topicTitle);
+        if (!descriptionList.isEmpty()) {
+            throw new CustomException(TOPIC_HAS_DESCRIPTION);
+        }
 
         topicRepository.delete(topic);
         return true;
@@ -94,6 +119,16 @@ public class TopicService {
     public List<String> mergeKeywordList(String keywords) {
         String[] split = keywords.split(",");
         return Arrays.asList(split);
+    }
+
+    private void addDupDates(Topic topic) {
+        List<Topic> topicList = dupDateRepository.queryTopicsByDupDate(topic.getStartDate(), topic.getEndDate());
+        List<DupDate> dupDateList = new ArrayList<>();
+
+        for (Topic descriptionTopic : topicList) {
+            dupDateList.add(new DupDate(topic, descriptionTopic));
+        }
+        dupDateRepository.saveAll(dupDateList);
     }
 
     private Chapter checkChapter(int num) {
